update Role Claim, move user controller tests to a new file
